fix: rename multichannel components, extension from .js to .jsx
